Random patches and comments (#5)
* Get rid of Clap: it's slow to compile and bloats the app

* cleanup panics

* invert settings interface condition

to avoid expect()

* add config note

* cleanup wireguard config

* cleanups and comments

* Fixes according review

* Prep for release

---------
