+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pony"
-version = "0.1.21"
+version = "0.1.22"
 edition = "2021"
 build = "build.rs" 
 
@@ -19,10 +19,10 @@ anyhow = "1"
 async-trait = "0.1"
 base64 = "0.22"
 chrono = { version = "0.4", features = ["serde"] }
-clap = { version = "4.0", features = ["derive"] }
 config = "0.14"
 clickhouse = { version = "0.13.0" }
 defguard_wireguard_rs = {version = "0.7.2", features=["serde"]}
+default-net = "0.22"
 fern = "0.7"
 futures = "0.3"
 ipnet = { version = "2", features=["serde"] }
@@ -46,13 +46,13 @@ tokio = { version = "1", features = ["full"] }
 tokio-postgres = { version="0.7", features=["with-uuid-1", "with-chrono-0_4", "with-serde_json-1"]}
 postgres-types = { version = "0.2", features = ["derive"] }
 uuid = { version = "1", features = ["serde", "v4"]  }
+urlencoding = "2.1.3"
+qrcode = "0.14"
 zmq = "0.10"
-default-net = "0.22"
 warp = "0.3"
 url = "2"
 x25519-dalek =  { version = "2", features = ["static_secrets"] }
 console-subscriber = {version = "0.4", optional = true} 
-urlencoding = "2.1.3"
 
 
 [build-dependencies]
