+++ Dockerfile.api
@@ -10,5 +10,5 @@ FROM rust:latest
 COPY --from=builder /app/target/release/api /app/api
 COPY --from=builder /app/config-api-example.toml /app/config.toml
 WORKDIR /app
-CMD ["api", "--config", "config.toml"]
+CMD ["api", "config.toml"]
