+++ src/bin/api/main.rs
@@ -1,5 +1,4 @@
 use chrono::NaiveTime;
-use clap::Parser;
 use fern::Dispatch;
 use futures::future::join_all;
 use std::net::Ipv4Addr;
@@ -29,27 +28,19 @@ use crate::core::ApiState;
 
 mod core;
 
-#[derive(Parser)]
-#[command(about = "Pony Api - control tool for Xray/Wireguard")]
-struct Cli {
-    #[arg(short, long, default_value = "config.toml")]
-    config: String,
-}
-
 #[tokio::main]
 async fn main() -> Result<()> {
     #[cfg(feature = "debug")]
     console_subscriber::init();
 
-    let args = Cli::parse();
+    let config_path = &std::env::args()
+        .nth(1)
+        .expect("required config path as an argument");
+    println!("Config file {:?}", config_path);
 
-    println!("Config file {:?}", args.config);
+    let mut settings = ApiSettings::new(&config_path);
 
-    let mut settings = ApiSettings::new(&args.config);
-
-    if let Err(e) = settings.validate() {
-        panic!("Wrong settings file {}", e);
-    }
+    settings.validate().expect("Wrong settings file");
     println!(">>> Settings: {:?}", settings.clone());
 
     // Logs handler init
@@ -70,10 +61,7 @@ async fn main() -> Result<()> {
 
     let debug = settings.debug.enabled;
 
-    let db = match PgContext::init(&settings.pg).await {
-        Ok(ctx) => ctx,
-        Err(e) => panic!("DB error: {}", e),
-    };
+    let db = PgContext::init(&settings.pg).await.expect("DB error");
 
     let ch = ChContext::new(&settings.clickhouse.address);
     let publisher = ZmqPublisher::new(&settings.zmq.endpoint).await;
