+++ src/state/node.rs
@@ -74,12 +74,12 @@ impl Node {
         let now = Utc::now();
         let mut inbounds: HashMap<Tag, Inbound> = HashMap::new();
 
-        let _ = {
+        {
             if let Some(config) = xray_config {
                 let xray_inbounds = config
                     .inbounds
                     .into_iter()
-                    .map(|inbound| (inbound.tag.clone(), inbound))
+                    .map(|inbound| (inbound.tag, inbound))
                     .collect::<HashMap<Tag, Inbound>>();
 
                 inbounds.extend(xray_inbounds);
@@ -104,12 +104,12 @@ impl Node {
         Self {
             uuid: settings.uuid,
             env: settings.env,
-            hostname: settings.hostname.expect("hostname"),
+            hostname: settings.hostname,
             status: Status::Online,
-            address: settings.address.expect("address"),
+            address: settings.address,
             created_at: now,
             label: settings.label,
-            interface: settings.default_interface.expect("default_interface"),
+            interface: settings.default_interface,
             modified_at: now,
             inbounds: inbounds,
         }
@@ -127,7 +127,7 @@ impl Node {
             env: self.env.clone(),
             hostname: self.hostname.clone(),
             address: self.address,
-            uuid: self.uuid.clone(),
+            uuid: self.uuid,
             inbounds: inbound_response,
             status: self.status,
             label: self.label.clone(),
