+++ src/zmq/mod.rs
@@ -12,7 +12,7 @@ pub enum Topic {
 
 impl Topic {
     pub fn from_raw(raw: &str) -> Self {
-        if let Ok(_) = uuid::Uuid::parse_str(raw) {
+        if uuid::Uuid::parse_str(raw).is_ok() {
             Topic::Init(raw.to_string())
         } else if raw == "all" {
             Topic::All
@@ -23,14 +23,14 @@ impl Topic {
 
     pub fn as_zmq_topic(&self) -> String {
         match self {
-            Topic::Updates(env) => format!("{env}"),
-            Topic::Init(uuid) => format!("{uuid}"),
+            Topic::Updates(env) => env.into(),
+            Topic::Init(uuid) => uuid.into(),
             Topic::Unknown(s) => s.clone(),
             Topic::All => "all".to_string(),
         }
     }
 
     pub fn all(uuid: &uuid::Uuid, env: &str) -> Vec<String> {
-        vec![format!("{uuid}"), format!("{env}"), "all".to_string()]
+        vec![uuid.to_string(), env.to_string(), "all".to_string()]
     }
 }
